updated logout Exception structure. Now throws ResponseException. Added check for auth being null.

+++ server/src/main/java/service/LogoutService.java
@@ -1,7 +1,7 @@
 package service;
 
 import dataaccess.AuthDAO;
-import dataaccess.DataAccessException;
+import exception.ResponseException;
 
 public class LogoutService {
     private AuthDAO authDAO;
@@ -10,10 +10,10 @@ public class LogoutService {
         this.authDAO = authDAO;
     }
 
-    public void logoutUser(String authToken) throws DataAccessException {
+    public void logoutUser(String authToken) throws ResponseException {
         var auth = authDAO.getAuth(authToken);
-        if (auth.username() == null) {
-            throw new DataAccessException("No matching user");
+        if (auth == null || auth.username() == null) {
+            throw new ResponseException(ResponseException.Code.Unauthorized, "Unauthorized");
         }
         authDAO.deleteAuth(authToken);
     }
